+++ unit_tests/test_pig_latin_logic.py
@@ -1,5 +1,5 @@
 import pytest
-from pig_latin import PigLatin
+from pig_latin.pig_latin import PigLatin
 
 
 class TestPigLatin:
